backend/django/requirements.txt
Django==3.2.5
psycopg2==2.8.6
daphne
djangorestframework
channels
channels_redis